Split TextServicesClient into 2 interfaces
Avoids fat interface where both text builder and text search clients are
implemented as one, BackgroundHandler will call both but API will only
ever call 1. Also allows for separate control of Timeouts etc per client
