+++ src/IIIFPresentation/API/Features/Manifest/ManifestWriteService.cs
@@ -101,7 +101,7 @@ public class ManifestWriteService(
     IManifestStorageManager manifestStorageManager,
     IPathRewriteParser pathRewriteParser,
     ILockManager manifestLockManager,
-    ITextServicesClient textServicesClient,
+    ITextBuilderClient textServicesClient,
     IOptions<AWSSettings> awsOptions,
     ILogger<ManifestWriteService> logger) : IManifestWrite
 {
@@ -447,6 +447,7 @@ public class ManifestWriteService(
 
         // When there is further work to do the JSON saved to S3 differs substantially from the original payload,
         // and we will want to store it. Otherwise, we'll pass null not to store the raw request.
+        // TODO - does this need to take into account Pipelines, or returned to highlight it's DLCS content resources?
         var requiresExternalContent = hasAssets || hasAdjuncts;
 
         var originalToStore = requiresExternalContent ? request.RawRequestBody : null;
@@ -478,6 +479,10 @@ public class ManifestWriteService(
             await manifestStorageManager.SaveManifestInStorage(iiifManifest, dbManifest, originalToStore,
                 saveToStaging, cancellationToken);
 
+            // TODO - how does this work for pipelines?
+            // Direct save (built upfront, no external content) with nothing to store as original:
+            // remove any stale original payload left by a previous version of this manifest.
+            // if (originalToStore is null)
             if (!saveToStaging && originalToStore is null)
             {
                 await manifestStorageManager.DeleteOriginalPayload(dbManifest);
@@ -528,7 +533,7 @@ public class ManifestWriteService(
         return null;
     }
 
-    private PipelineJob? BuildPipelineJob(DbManifest dbManifest, List<PipelineItem> pipeline)
+    private static PipelineJob? BuildPipelineJob(DbManifest dbManifest, List<PipelineItem> pipeline)
     {
         // Returns a job for the first recognised pipeline step; additional steps of the same type are ignored.
         foreach (var pipelineItem in pipeline)
