+++ src/IIIFPresentation/BackgroundHandler/Program.cs
@@ -39,7 +39,8 @@ builder.RegisterSharedServiceSettings();
 builder.Services.AddAws(builder.Configuration, builder.Environment)
     .AddDataAccess(builder.Configuration)
     .AddDlcsOrchestratorClient(dlcs)
-    .AddTextServicesClient()
+    .AddTextBuilderClient()
+    .AddTextAugmentedManifestClient()
     .AddBackgroundServices(aws)
     .AddSingleton<IPathGenerator, SettingsBasedPathGenerator>()
     .AddSingleton<SettingsBasedPathGenerator>()
