+++ src/IIIFPresentation/DLCS/ServiceCollectionX.cs
@@ -1,5 +1,4 @@
 using System.Net.Http.Headers;
-using System.Reflection;
 using System.Runtime.CompilerServices;
 using DLCS.API;
 using DLCS.Handlers;
